+++ variables.tf
@@ -3,17 +3,16 @@ variable "ami_id" {
   type    = string
 
   description = <<-EOT
-    an AMI ID.
-    https://go.s3d.club/tf/ec2#ami_id
+    An AMI ID.
     EOT
 }
 
 variable "ami_suffix" {
   default = "arm64-gp2"
+  type    = string
 
   description = <<-EOT
-    an AMI suffix that is used to find an ami when ami_id is null.
-    https://go.s3d.club/tf/ec2#ami_suffix
+    An AMI suffix that is used to find an ami when ami_id is null.
     EOT
 }
 
@@ -21,8 +20,7 @@ variable "ssh_cidrs" {
   type = list(string)
 
   description = <<-EOT
-    a list of addresses for ssh ingress.
-    https://go.s3d.club/tf/ec2#ssh_cidrs
+    A list of addresses for ssh ingress.
     EOT
 }
 
@@ -31,8 +29,7 @@ variable "ssh_cidr6s" {
   type    = list(string)
 
   description = <<-EOT
-    a list of addresses for ssh ingress.
-    https://go.s3d.club/tf/ec2#ssh_cidr6s
+    A list of addresses for ssh ingress.
     EOT
 }
 
@@ -40,8 +37,7 @@ variable "domain" {
   type = string
 
   description = <<-EOT
-    the domain for route53 registration.
-    https://go.s3d.club/tf/ec2#domain
+    The domain for route53 registration.
     EOT
 }
 
@@ -49,8 +45,7 @@ variable "egress_cidrs" {
   type = list(string)
 
   description = <<-EOT
-    a list of addresses for open egress.
-    https://go.s3d.club/tf/ec2#egress_cidrs
+    A list of addresses for open egress.
     EOT
 }
 
@@ -59,8 +54,7 @@ variable "egress_cidr6s" {
   type    = list(string)
 
   description = <<-EOT
-    a list of addresses for open egress.
-    https://go.s3d.club/tf/ec2#egress_cidr6s
+    A list of addresses for open egress.
     EOT
 }
 
@@ -69,8 +63,7 @@ variable "instance_type" {
   type    = string
 
   description = <<-EOT
-    an AWS EC2 instance type.
-    https://go.s3d.club/tf/ec2#instance_type
+    An AWS EC2 instance type.
     EOT
 }
 
@@ -78,18 +71,16 @@ variable "name_prefix" {
   type = string
 
   description = <<-EOT
-    an prefix for naming resources.
-    https://go.s3d.club/tf/ec2#name_prefix
+    An prefix for naming resources.
     EOT
 }
 
 variable "setup_ref" {
-  default = "v0.1.13" # from dev-setup
+  default = "v0.1.15" # from dev-setup
   type    = string
 
   description = <<-EOT
-    a version used when fetching the setup script.
-    https://go.s3d.club/tf/ec2#setup_ref
+    A version used when fetching the setup script.
     EOT
 }
 
@@ -97,8 +88,7 @@ variable "subnet_id" {
   type = string
 
   description = <<-EOT
-    a subnet_id.
-    https://go.s3d.club/tf/ec2#subnet_id
+    A subnet_id.
     EOT
 }
 
@@ -106,8 +96,7 @@ variable "tags" {
   type = map(string)
 
   description = <<-EOT
-    a map of tags for resources.
-    https://go.s3d.club/tf/ec2#tags
+    Aa map of tags for resources.
     EOT
 }
 
@@ -116,8 +105,7 @@ variable "template" {
   type    = string
 
   description = <<-EOT
-    a template to use for setting up the image.
-    https://go.s3d.club/tf/ec2#template
+    A template to use for setting up the image.
     EOT
 }
 
@@ -125,8 +113,7 @@ variable "key_name" {
   type = string
 
   description = <<-EOT
-    an EC2 keyname.
-    https://go.s3d.club/tf/ec2#key_name
+    An EC2 keyname.
     EOT
 }
 
@@ -135,8 +122,7 @@ variable "username" {
   type    = string
 
   description = <<-EOT
-    an username.
-    https://go.s3d.club/tf/ec2#username
+    A username.
     EOT
 }
 
@@ -145,8 +131,7 @@ variable "volume_size" {
   type    = number
 
   description = <<-EOT
-    an size in GiB for the volume.
-    https://go.s3d.club/tf/ec2#volume_size
+    An size in GiB for the volume.
     EOT
 }
 
@@ -154,7 +139,6 @@ variable "vpc_id" {
   type = string
 
   description = <<-EOT
-    a vpc id.
-    https://go.s3d.club/tf/ec2#vpc_id
+    A vpc id.
     EOT
 }